ai_player/requirements.txt
nvidia-nat==1.2.1
nvidia-nat-langchain==1.2.1
openai==1.101.0
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2